refactor: member 모듈 구성 및 Docker 설정
