+++ member-service/compose.yml
@@ -1,0 +1,80 @@
+
+services:
+  # MySQL 서비스
+  my-mysql:
+    image: mysql:latest
+    container_name: member-mysql
+    environment:
+      MYSQL_ROOT_PASSWORD: password123
+    ports:
+      - "3307:3306"
+    volumes:
+      - /Users/jeongjuhyeon/개발/docker-mysql/mysql_data_member:/var/lib/mysql
+    networks:
+      - coupon-network
+    restart: always
+    healthcheck:
+      test: [ "CMD", "mysqladmin", "ping", "-h", "localhost" ]
+      interval: 30s
+      retries: 5
+      start_period: 30s
+      timeout: 10s
+
+  mysqld-exporter:
+    container_name: mysqld_exporter
+    image: prom/mysqld-exporter
+    environment:
+      - DATA_SOURCE_NAME="wngus:password123@(my-mysql:3306)/"
+    command:
+      - "--mysqld.username=wngus:password123"
+      - "--mysqld.address=my-mysql:3306"
+    ports:
+      - "9104:9104"
+    depends_on:
+      - my-mysql
+    restart: unless-stopped
+    networks:
+      - coupon-network  # 추가된 부분
+      -
+  # Prometheus 서비스
+  prometheus:
+    image: prom/prometheus:latest
+    container_name: prometheus
+    volumes:
+      - ./prometheus.yml:/etc/prometheus/prometheus.yml
+    command:
+      - '--enable-feature=native-histograms'
+      - '--web.enable-remote-write-receiver'
+      - '--config.file=/etc/prometheus/prometheus.yml'
+    ports:
+      - "9090:9090"
+    networks:
+      - coupon-network
+
+  # Grafana 서비스
+  grafana:
+    image: grafana/grafana:latest
+    container_name: grafana
+    ports:
+      - "3000:3000"
+    networks:
+      - coupon-network
+
+networks:
+  coupon-network:
+    driver: bridge
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
