+++ member-service/src/main/java/com/example/memberserver/member/applicaion/outputport/MemberOutputPort.java
@@ -1,0 +1,14 @@
+package com.example.memberserver.member.applicaion.outputport;
+
+
+import com.example.memberserver.member.domain.Member;
+
+import java.util.Optional;
+
+public interface MemberOutputPort {
+    boolean checkEmail(String email);
+    Member save(Member member);
+    Optional<Member> findById(Long id);
+    Optional<Member> findByEmail(String email);
+    boolean existsById(Long memberId);
+}
