+++ member-service/src/main/java/com/example/memberserver/member/applicaion/usecase/AuthMember.java
@@ -1,0 +1,11 @@
+package com.example.memberserver.member.applicaion.usecase;
+
+import com.example.memberserver.member.domain.Member;
+import com.example.memberserver.member.framwork.web.request.MemberInfoDTO;
+import com.example.memberserver.member.framwork.web.request.MemberLoginDTO;
+import com.example.memberserver.member.framwork.web.response.MemberLoginOutputDTO;
+
+public interface AuthMember {
+    MemberLoginOutputDTO login(MemberLoginDTO request);
+    Member signUp(MemberInfoDTO request);
+}
