+++ member-service/src/main/java/com/example/memberserver/member/auth/handler/CustomAccessDeniedHandler.java
@@ -1,0 +1,26 @@
+package com.example.memberserver.member.auth.handler;
+
+import com.example.common.global.exception.errorcode.CommonErrorCode;
+import com.example.memberserver.member.auth.jwt.JwtAuthenticationFilter;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.web.access.AccessDeniedHandler;
+
+import java.io.IOException;
+
+@Slf4j
+@RequiredArgsConstructor
+public class CustomAccessDeniedHandler implements AccessDeniedHandler {
+    /***
+     * '인가'가 실패했을 때의 메서드.
+     */
+    @Override
+    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException {
+        log.error("[ACCESS DENIED]");
+        JwtAuthenticationFilter.setErrorResponse(response, CommonErrorCode.JWT_AUTHORIZATION_FAILED);
+    }
+}
