+++ member-service/src/main/java/com/example/memberserver/member/domain/enumeration/Role.java
@@ -1,0 +1,14 @@
+package com.example.memberserver.member.domain.enumeration;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum Role {
+
+    ADMIN("ROLE_ADMIN"), USER("ROLE_USER");
+
+    private final String key;
+
+}
