+++ member-service/src/main/java/com/example/memberserver/member/infra/persistence/MemberJpaRepository.java
@@ -1,0 +1,14 @@
+package com.example.memberserver.member.infra.persistence;
+
+
+import com.example.memberserver.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberJpaRepository extends JpaRepository<Member,Long> {
+
+    boolean existsByEmail(String email);
+
+    Optional<Member> findByEmail(String email);
+}
