--- src/main/java/com/personal_project/coupon/member/domain/enumeration/Role.java
@@ -1,14 +1,0 @@
-package com.personal_project.coupon.member.domain.enumeration;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum Role {
-
-    ADMIN("ROLE_ADMIN"), USER("ROLE_USER");
-
-    private final String key;
-
-}
