--- src/main/java/com/personal_project/coupon/member/framwork/web/request/MemberInfoDTO.java
@@ -1,24 +1,0 @@
-package com.personal_project.coupon.member.framwork.web.request;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
-import lombok.Getter;
-
-@Getter
-public class MemberInfoDTO {
-
-    @Pattern(regexp = "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+.[A-Za-z]{2,6}$", message = "이메일 형식이 올바르지 않습니다.")
-    @NotBlank(message = "이메일은 필수 입력 값입니다.")
-    private String email;
-
-    @NotBlank(message = "비밀번호는 필수 입력 값입니다.")
-    @Pattern(regexp = "(?=.*[0-9])(?=.*[a-zA-Z])(?=.*\\W)(?=\\S+$).{8,16}", message = "비밀번호는 8~16자 영문 대 소문자, 숫자, 특수문자를 사용하세요.")
-    private String password;
-
-    @NotBlank(message = "이름은 필수 입력 값입니다.")
-    private String name;
-
-    @NotBlank(message = "휴대폰 번호는 필수 입력 값입니다.")
-    @Pattern(regexp = "010-\\d{4}-\\d{4}", message = "휴대폰 번호는 010-XXXX-XXXX 형식으로 입력해주세요.")
-    private String phone;
-}
