--- src/main/java/com/personal_project/coupon/member/framwork/web/request/MemberLoginDTO.java
@@ -1,15 +1,0 @@
-package com.personal_project.coupon.member.framwork.web.request;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
-import lombok.Getter;
-
-@Getter
-public class MemberLoginDTO {
-    @Pattern(regexp = "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+.[A-Za-z]{2,6}$", message = "이메일 형식이 올바르지 않습니다.")
-    @NotBlank(message = "이메일은 필수 입력 값입니다.")
-    private String email;
-
-    @NotBlank(message = "비밀번호는 필수 입력 값입니다.")
-    private String password;
-}
